requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
flask-cors==3.0.10
requests==2.28.2
python-dotenv==0.21.1
mnemonic==0.20
bech32==1.1.0
gunicorn==20.1.0
